Try reuse existing window to display result
When displaying result in "other window", result buffer needs to be quit
first because it has been displayed itself in an other window. This
improved navigability when the target buffer is already displayed in
another window.

A previously failing case is such as :
|---------------+-------|
| file1<cursor> | file2 |
|---------------+-------|
| file3 | file4 |
|---------------+-------|

A rscope request is made which result in a unique match in file3:666
|-------+-------------------|
| file1 | file3:666<cursor> |
|-------+-------------------|
| file3 | file4 |
|-------+-------------------|

In this case, 2 buffers with file3 within are displayed, while it would be
better to have file3 buffer reused.

This patch changes the behaviour to reuse file3 buffer (bottom left).

Signed-off-by: Sylvain Chouleur <sylvain.chouleur@gmail.com>
Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>