Display single result in same window
In case of single result for a search, the entry is automatically
displayed without displaying the result buffer.

The previous behavior was to display the entry in an other window. This
can be annoying when working with only two windows, for example one for
code, one for documentation, we would like to keep always the same one
for code.

This patch changes this behavior to use the window from where we issued
the research to display the single result.

Signed-off-by: Sylvain Chouleur <sylvain.chouleur@gmail.com>
Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>