rscope: add cscope process auto recover
When working on remote servers and switching networks, the cscope
process dies more often than when working locally. This patch makes
the `rscope-handle-query' function ask the end-user if he wants to run
`rscope-init' to recover the situation. This patch also introduces a
`rscope-auto-reinit' customization variable to enforce the automatic
call of `rscope-init'.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>